+++ mod-record-specifications-server/src/main/java/org/folio/rspec/service/processor/request/strategy/UpdateRequestProcessingStrategy.java
@@ -1,0 +1,11 @@
+package org.folio.rspec.service.processor.request.strategy;
+
+import org.folio.rspec.domain.dto.DefinitionType;
+import org.folio.rspec.domain.dto.UpdateRequestEvent;
+
+public interface UpdateRequestProcessingStrategy {
+
+  void process(UpdateRequestEvent event);
+
+  DefinitionType getType();
+}